Refactoring to reorganise services
Note this could end up in a different structure
