+++ app/controllers/api/v1/moves_controller.rb
@@ -4,19 +4,15 @@ module Api
   module V1
     class MovesController < ApplicationController
       def index
-        render json: MoveFinder.new(filter_params).scope
+        render json: Moves::MoveFinder.new(filter_params).call
       end
 
       private
 
+      PERMITTED_FILTER_PARAMS = %i[date_from date_to from_location_id location_type status].freeze
+
       def filter_params
-        params.fetch(:filter, {}).permit(
-          :date_from,
-          :date_to,
-          :from_location_id,
-          :location_type,
-          :status
-        )
+        params.fetch(:filter, {}).permit(PERMITTED_FILTER_PARAMS)
       end
     end
   end
